Refactor: RedisUtil 클래스 생성
