+++ core/src/main/java/org/bookwoori/core/domain/member/facade/AuthFacade.java
@@ -12,7 +12,7 @@ import org.bookwoori.core.domain.serverMember.entity.ServerRole;
 import org.bookwoori.core.domain.serverMember.service.ServerMemberServiceImpl;
 import org.bookwoori.core.global.exception.CustomException;
 import org.bookwoori.core.global.exception.ErrorCode;
-import org.bookwoori.core.global.s3.S3Util;
+import org.bookwoori.core.global.util.S3Util;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
