+++ core/src/main/java/org/bookwoori/core/global/util/RedisUtil.java
@@ -1,0 +1,23 @@
+package org.bookwoori.core.global.util;
+
+import java.time.Duration;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+@Component
+@RequiredArgsConstructor
+public class RedisUtil {
+
+    private final RedisTemplate<String, Object> redisTemplate;
+
+    public void setValuesWithTimeout(String key, String value, Duration timeout) {
+        redisTemplate.opsForValue().set(key, value, timeout);
+    }
+
+    @Transactional(readOnly = true)
+    public Object getValues(String key) {
+        return redisTemplate.opsForValue().get(key);
+    }
+}
