+++ core/src/main/java/org/bookwoori/core/global/util/S3Util.java
@@ -1,4 +1,4 @@
-package org.bookwoori.core.global.s3;
+package org.bookwoori.core.global.util;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
